moved Navbar to layout
