+++ client/src/App.js
@@ -2,7 +2,7 @@ import React, { Fragment, useEffect } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Landing from './components/layout/landing/Landing';
 import Routes from './components/routing/Routes';
-import Navbar from './components/elements/navbar/Navbar';
+import Navbar from './components/layout/navbar/Navbar';
 
 
 import './App.css';
